+++ dst/routes/accounts.js
@@ -10,7 +10,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * 口座ルーター
- * @ignore
  */
 const pecorinoapi = require("@pecorino/api-nodejs-client");
 const createDebug = require("debug");
@@ -20,25 +19,66 @@ const accountsRouter = express.Router();
 /**
  * 口座検索
  */
-accountsRouter.get('/', (req, res, next) => __awaiter(this, void 0, void 0, function* () {
+accountsRouter.get('/', 
+// tslint:disable-next-line:cyclomatic-complexity
+(req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
         const accountService = new pecorinoapi.service.Account({
             endpoint: process.env.PECORINO_API_ENDPOINT,
             auth: req.user.authClient
         });
-        debug('searching accounts...', req.query);
-        const accounts = yield accountService.search({
+        const searchConditions = {
+            limit: req.query.limit,
+            page: req.query.page,
             accountType: req.query.accountType,
             accountNumbers: (typeof req.query.accountNumber === 'string' && req.query.accountNumber.length > 0) ?
                 [req.query.accountNumber] :
                 [],
             statuses: [],
-            name: req.query.name,
-            limit: 100
+            name: req.query.name
+        };
+        if (req.query.format === 'datatable') {
+            debug('searching accounts...', req.query);
+            const { totalCount, data } = yield accountService.searchWithTotalCount(searchConditions);
+            res.json({
+                draw: req.query.draw,
+                recordsTotal: totalCount,
+                recordsFiltered: totalCount,
+                data: data
+            });
+        }
+        else {
+            res.render('accounts/index', {
+                query: req.query
+            });
+        }
+    }
+    catch (error) {
+        next(error);
+    }
+}));
+/**
+ * 口座詳細
+ */
+accountsRouter.get('/:accountType/:accountNumber', (req, res, next) => __awaiter(this, void 0, void 0, function* () {
+    try {
+        const accountService = new pecorinoapi.service.Account({
+            endpoint: process.env.PECORINO_API_ENDPOINT,
+            auth: req.user.authClient
         });
-        res.render('accounts/index', {
-            query: req.query,
-            accounts: accounts
+        const { totalCount, data } = yield accountService.searchWithTotalCount({
+            limit: 1,
+            statuses: [],
+            accountType: req.params.accountType,
+            accountNumbers: [req.params.accountNumber]
+        });
+        if (totalCount < 1) {
+            throw new pecorinoapi.factory.errors.NotFound('Account');
+        }
+        const account = data[0];
+        res.render('accounts/show', {
+            message: '',
+            account: account
         });
     }
     catch (error) {
@@ -46,23 +86,24 @@ accountsRouter.get('/', (req, res, next) => __awaiter(this, void 0, void 0, func
     }
 }));
 /**
- * 口座に対する転送アクション検索
+ * 取引検索
  */
-accountsRouter.get('/:accountType/:accountNumber/actions/MoneyTransfer', (req, res, next) => __awaiter(this, void 0, void 0, function* () {
+accountsRouter.get('/:accountType/:accountNumber/actions/moneyTransfer', (req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
         const accountService = new pecorinoapi.service.Account({
             endpoint: process.env.PECORINO_API_ENDPOINT,
             auth: req.user.authClient
         });
-        const actions = yield accountService.searchMoneyTransferActions({
-            accountType: req.params.accountType,
-            accountNumber: req.params.accountNumber
-        });
-        res.render('accounts/actions/moneyTransfer', {
+        const searchActionsResult = yield accountService.searchMoneyTransferActionsWithTotalCount({
+            limit: req.query.limit,
+            page: req.query.page,
             accountType: req.params.accountType,
             accountNumber: req.params.accountNumber,
-            actions: actions
+            sort: {
+                endDate: pecorinoapi.factory.sortType.Descending
+            }
         });
+        res.json(searchActionsResult);
     }
     catch (error) {
         next(error);