+++ dst/routes/transactions.js
@@ -10,7 +10,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * 取引ルーター
- * @ignore
  */
 const pecorinoapi = require("@pecorino/api-nodejs-client");
 const createDebug = require("debug");
@@ -117,17 +116,16 @@ transactionsRouter.all('/deposit/:transactionId/confirm', (req, res, next) => __
                 endpoint: process.env.PECORINO_API_ENDPOINT,
                 auth: req.user.authClient
             });
-            const accounts = yield accountService.search({
-                accountType: transaction.object.accountType,
-                accountNumbers: [transaction.object.toAccountNumber],
+            const searchAccountsResult = yield accountService.searchWithTotalCount({
+                accountType: transaction.object.toLocation.accountType,
+                accountNumbers: [transaction.object.toLocation.accountNumber],
                 statuses: [],
                 limit: 1
             });
-            const account = accounts.shift();
-            if (account === undefined) {
-                throw new Error('to account not found');
+            toAccount = searchAccountsResult.data.shift();
+            if (toAccount === undefined) {
+                throw new Error('To Location Not Found');
             }
-            toAccount = account;
         }
         res.render('transactions/deposit/confirm', {
             transaction: transaction,
@@ -225,17 +223,16 @@ transactionsRouter.all('/withdraw/:transactionId/confirm', (req, res, next) => _
                 endpoint: process.env.PECORINO_API_ENDPOINT,
                 auth: req.user.authClient
             });
-            const accounts = yield accountService.search({
-                accountType: transaction.object.accountType,
-                accountNumbers: [transaction.object.fromAccountNumber],
+            const searchAccountsResult = yield accountService.searchWithTotalCount({
+                accountType: transaction.object.fromLocation.accountType,
+                accountNumbers: [transaction.object.fromLocation.accountNumber],
                 statuses: [],
                 limit: 1
             });
-            const account = accounts.shift();
-            if (account === undefined) {
-                throw new Error('to account not found');
+            fromAccount = searchAccountsResult.data.shift();
+            if (fromAccount === undefined) {
+                throw new Error('From Location Not Found');
             }
-            fromAccount = account;
         }
         res.render('transactions/withdraw/confirm', {
             transaction: transaction,