+++ CHANGELOG.md
@@ -18,6 +18,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v1.1.0 - 2019-02-19
+
+### Added
+
+- 口座検索と取引履歴検索にページネーション追加
+
+### Changed
+
+- 口座詳細コンテンツを追加
+- ログイン時にstateを継承するように変更
+- install @pecorino/api-nodeje-client@2.0.0
+
 ## v1.0.0 - 2018-08-01
 
 ### Changed